moving the SolverResults object to its own source file
