+++ doc/reference/solver_results.rst
@@ -1,0 +1,5 @@
+pybnb.solver_results
+--------------------
+
+.. automodule:: pybnb.solver_results
+    :members:
